fix Transform & Deliver bug (#29)
